requirements.txt
onnxruntime
numpy
Pillow